+++ README.md
@@ -53,27 +53,22 @@ PORT - желаемый порт, например 800
 
 # Kitoboy Service
 
-Структура проекта: 
 
-Для поднятия контейнера с backend'ом необходимо выполнить:
+Для поднятия контейнера с сервисом kitoboy необходимо выполнить последовательность команд:
 
-1 sudo docker-compose up -d --build    # Поднятие контейнера
-2 sudo docker exec kitoboy python3.6 db_manager.py create_db   # Создание таблиц в докере
-3 sudo docker exec kitoboy ./scripts/./db_fill.sh dev | prod     # наполнение таблиц тестовыми данными:
+1 docker-compose up -d --build # Поднятие контейнера
+2 docker exec app python3.8 db_manager.py drop_db   # Удаление таблиц в бд (опционально)
+3 docker exec app python3.8 db_manager.py create_db  # Создание таблиц описанных в models (init data base)
+3 docker exec app ./scripts/./db_fill.sh dev    # наполнение таблиц тестовыми данными
 
 check on:
 localhost:8000
 
 
-Для входа в бд в контейнере:
-1 sudo docker exec -it db psql -U test  -d  kitoboy_db  --password
+Для входа в контейнере базы данных:
+1 docker exec -it db psql -U test  -d kitoboy_db  --password
 
-
-#3. sudo docker-compose -f docker-compose-db.yaml restart (применить pg_hba.conf)
-#Чтобы изменить environment - нужно 1ую команду заменить на :
-#sudo docker-compose -f docker-compose-db.yaml --env-file PATH up --build
-
-
+(базовые команды для субд PostgreSQL см. https://www.postgresqltutorial.com/postgresql-administration/psql-commands/) 
 
 
 # Migration Managment
