+++ app/templates/index.html
@@ -6,6 +6,6 @@
     <title>Flask Docker </title>
 </head>
 <body>
-    <h1>This is a Flask App containerised with Docker</h1>
+    <h1>This is a Flask App containerised with Docker!</h1>
 </body>
 </html>
