--- config.py
@@ -1,16 +1,0 @@
-import os
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-class Config(object):
-    # ...
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
-        'sqlite:///' + os.path.join(basedir, 'siniy_kit.db')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    DEBUG = True
-    SECRET_KEY = 'development key'
-    USERNAME = 'admin'
-    PASSWORD = 'default'
-    UPLOAD_FOLDER = '/tmp/'
-    ALLOWED_EXTENSIONS = set(['txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif', 'json'])
-    POSTS_PER_PAGE = 10
-    SUICIDAL_RATING = [(0, "Normal"), (1, "Sus"), (2, "High")]
