+++ dev_config.py
@@ -16,8 +16,8 @@ class DevelopmentConfig(Config):
     COMMON_DATABASE_PATH = f'postgresql://{DB_USER}:{DB_PASS}@{DB_SERVICE}:{DB_PORT}'
     SQLALCHEMY_DATABASE_URI = f'{COMMON_DATABASE_PATH}/{DB_NAME}'
 
-    CELERY_BROKER_URL = os.environ.get('CELERY_BROKER_URL',
-                                       'redis://localhost:6379'),
-    CELERY_RESULT_BACKEND = os.environ.get('CELERY_RESULT_BACKEND',
-                                           'redis://localhost:6379')
+    # CELERY_BROKER_URL = os.environ.get('CELERY_BROKER_URL',
+    #                                    'redis://localhost:6379'),
+    # CELERY_RESULT_BACKEND = os.environ.get('CELERY_RESULT_BACKEND',
+    #                                        'redis://localhost:6379')
