--- draft.py
@@ -1,1 +1,0 @@
-from flask_script import Manager, MigrateCommand
