+++ env.env
@@ -1,7 +1,7 @@
 POSTGRES_USER=test
 POSTGRES_DB=kitoboy_db
 POSTGRES_PASSWORD=12345
-POSTGRES_SERVICE=kitoboy_db
+POSTGRES_SERVICE=db
 POSTGRES_PORT=5432
 CELERY_BROKER_URL=redis://redis_backend:6379
 CELERY_RESULT_BACKEND=redis://redis_backend:6379
