--- postgres_extra/data_base.env
@@ -1,3 +1,0 @@
-POSTGRES_USER=test
-POSTGRES_DB=kitoboy_db
-POSTGRES_PASSWORD=12345
