+++ requirements.txt
@@ -1,27 +1,73 @@
-flask
-selenium==3.141.0
-dostoevsky==0.6.0
-pandas
-pymorphy2==0.9.1
-nltk==3.6.5
-numpy
-matplotlib==2.2.3
-gunicorn==20.1.0
-Telethon==1.24.0
-Flask-SQLAlchemy
-Flask-Migrate==2.5.2
-Flask-Login
-Flask-Bcrypt
-Werkzeug
-psycopg2-binary==2.8.4
-flask-restplus
-Flask-Cors
-flask-marshmallow
-marshmallow
-marshmallow-sqlalchemy
-flask_jwt_extended
-flask_script==2.0.3
-sqlalchemy
-flask-caching
+alembic==1.8.1
+amqp==2.6.1
+aniso8601==9.0.1
+attrs==22.1.0
+bcrypt==3.2.2
+billiard==3.6.4.0
+cachelib==0.9.0
 celery==4.4.5
+cffi==1.15.1
+click==8.1.3
+cycler==0.11.0
+DAWG-Python==0.7.2
+docopt==0.6.2
+dostoevsky==0.6.0
+fasttext==0.9.2
+Flask==2.1.2
+Flask-Bcrypt==1.0.1
+Flask-Caching==2.0.0
+Flask-Cors==3.0.10
+Flask-JWT-Extended==4.4.3
+Flask-Login==0.6.2
+flask-marshmallow==0.14.0
+Flask-Migrate==2.5.2
+flask-restplus==0.13.0
+Flask-Script==2.0.3
+Flask-SQLAlchemy==2.5.1
+fonttools==4.34.4
+future==0.18.2
+greenlet==1.1.2
+gunicorn==20.1.0
+importlib-metadata==4.12.0
+importlib-resources==5.9.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+joblib==1.1.0
+jsonschema==4.8.0
+kiwisolver==1.4.4
+kombu==4.6.11
+Mako==1.2.1
+MarkupSafe==2.1.1
+marshmallow==3.17.0
+marshmallow-sqlalchemy==0.28.1
+matplotlib==3.5.2
+nltk==3.7
+numpy==1.23.1
+packaging==21.3
+pandas==1.4.3
+Pillow==9.2.0
+psycopg2-binary==2.8.4
+pyaes==1.6.1
+pyasn1==0.4.8
+pybind11==2.10.0
+pycparser==2.21
+PyJWT==2.4.0
+pymorphy2==0.9.1
+pymorphy2-dicts-ru==2.4.417127.4579844
+pyparsing==3.0.9
+pyrsistent==0.18.1
+python-dateutil==2.8.2
+pytz==2022.1
+razdel==0.5.0
 redis==3.5.3
+regex==2022.7.25
+rsa==4.9
+selenium==3.141.0
+six==1.16.0
+SQLAlchemy==1.4.39
+Telethon==1.24.0
+tqdm==4.64.0
+urllib3==1.26.11
+vine==1.3.0
+Werkzeug==2.2.1
+zipp==3.8.1
