+++ scripts/db_fill.sh
@@ -6,8 +6,8 @@
 if [ "$1" ]
 then
   mode="$1"
-  python3.6 $WORK_DIR/db_manager.py create_digital_users -f $WORK_DIR/data_for_db/digital_user.json
-  python3.6 $WORK_DIR/db_manager.py create_posts
+  python3.8 $WORK_DIR/db_manager.py create_digital_users -f $WORK_DIR/data_for_db/digital_user.json
+  python3.8 $WORK_DIR/db_manager.py create_posts
   echo Done
