+++ supervisord.conf
@@ -4,28 +4,19 @@ umask=022
 
 
 [program:gunicorn]
-directory=/var/www/labeling_service/backend
+directory=/var/www/kitoboy_service/backend
 command=gunicorn -w 6 -b 0.0.0.0:8000 main:app
 autostart=true
 autorestart=true
-
-
-[program:celery_tasks]
-process_name=%(program_name)
-directory=/var/www/labeling_service/backend
-command=/usr/local/bin/celery -A celery_tasks.celery worker -E --loglevel=info
-stdout_logfile=/var/www/labeling_service/backend/app/logs/%(program_name)s-stdout.log
-stderr_logfile=/var/www/labeling_service/backend/app/logs/%(program_name)s-stderr.log
-user=celery
-logfile_maxbytes=50MB
-loglevel=info
-autostart=true
-autorestart=true
-
 
 
 [program:nginx]
 command=/usr/sbin/nginx
 autostart=true
 autorestart=true
+
+
+
+
+
