--- wsgi.py
@@ -1,4 +1,0 @@
-from main_2 import app
-
-if __name__ == "__main__":
-    app.run()
